refactor(agent-tars-cli): decouple 5
