+++ multimodal/agent-tars-cli/src/core/workspace.ts
